+++ metrics/prometheus/prometheus.go
@@ -103,12 +103,12 @@ func (g prometheusGauge) With(f metrics.Field) metrics.Gauge {
 	}
 }
 
-func (g prometheusGauge) Set(value int64) {
-	g.GaugeVec.With(prometheus.Labels(g.Pairs)).Set(float64(value))
+func (g prometheusGauge) Set(value float64) {
+	g.GaugeVec.With(prometheus.Labels(g.Pairs)).Set(value)
 }
 
-func (g prometheusGauge) Add(delta int64) {
-	g.GaugeVec.With(prometheus.Labels(g.Pairs)).Add(float64(delta))
+func (g prometheusGauge) Add(delta float64) {
+	g.GaugeVec.With(prometheus.Labels(g.Pairs)).Add(delta)
 }
 
 type prometheusHistogram struct {
